% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{fipe}
\alias{fipe}
\title{Consulta o valor do carro na tabela FIPE}
\usage{
fipe(cod_ref, cod_marca, cod_modelo, cod_ano)
}
\description{
Consulta o valor do carro a partir dos parametros necessarios para a
caracterizacao completa do modelo
}
