% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{fipe_ano}
\alias{fipe_ano}
\title{Anos dos modelos disponiveis na tabela FIPE}
\usage{
fipe_ano(cod_ref = NULL, cod_marca, cod_modelo)
}
\description{
Consulta o ano do modelo disponiveis para um determinado mes de referencia,
marca e modelo
}
