% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fipe.R
\name{fipe_carro}
\alias{fipe_carro}
\title{Consulta o valor do carro disponiveis na Tabela FIPE}
\usage{
fipe_carro(modelo, marca = NULL, ano = NULL,
  data_referencia = Sys.Date(), .progress = FALSE)
}
\arguments{
\item{modelo}{vetor de caracteres. Nome de um ou mais carros que se deseja consultar.}

\item{marca}{vetor de caracteres. Se NULL, consulta em todas o modelo em todas as marcas, do contrario apenas nas indicadas.}

\item{ano}{vetor numerico. Indica um ou mais anos de fabricacao a serem consultados. Valor 0 indica carro 0 km.}

\item{data_referencia}{vetor de datas. Indica a data de referencia que o valor carro foi consultado.}
}
\value{
Um tibble com resultado da consulta.
}
\description{
Consulta o valor do carro disponiveis na Tabela FIPE
}
\examples{
fipe_carro("etios", "toyota", c(2018, 0), c("2019-01-01", "2018-01-01", "2017-01-01"))

}
