% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{fipe_modelo}
\alias{fipe_modelo}
\title{Modelos disponiveis na tabela FIPE}
\usage{
fipe_modelo(cod_ref = NULL, cod_marca)
}
\description{
Consulta os modelos disponiveis para um determinado mes de referencia e marca
}
