% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fipe.R
\name{fipe_vehicle}
\alias{fipe_vehicle}
\title{Consulta o price do carro disponiveis na Tabela FIPE}
\usage{
fipe_vehicle(model, make = NULL, year = NULL, date = Sys.Date(),
  .progress = FALSE)
}
\arguments{
\item{model}{vetor de caracteres. Nome de um ou mais carros que se deseja consultar.}

\item{make}{vetor de caracteres. Se NULL, consulta em todas o modelo em todas as marcas, do contrario apenas nas indicadas.}

\item{year}{vetor numerico. Indica um ou mais anos de fabricacao a serem consultados. price 0 indica carro 0 km.}

\item{date}{vetor de datas. Indica a data de referencia que o price carro foi consultado.}

\item{.progress}{a logical, for whether or not to print a progress bar.}
}
\value{
Um tibble com resultado da consulta.
}
\description{
Consulta o price do carro disponiveis na Tabela FIPE
}
\examples{
fipe_vehicle(model = "etios", make = "toyota", year = 0)

}
