% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_fipe.R
\name{get_marcas}
\alias{get_marcas}
\title{Codigo das marcas}
\usage{
get_marcas(referencia)
}
\arguments{
\item{referencia}{Codigo do ano/mes de referencia.
Ver \code{\link{get_referencia}}}
}
\value{
Uma tibble com codigo das marcas
}
\description{
Codigo das marcas
}
\examples{
get_marcas(get_referencia(2022))
}
