% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_fipe.R
\name{get_modelos}
\alias{get_modelos}
\title{Codigo dos modelos}
\usage{
get_modelos(referencia, cod_marca)
}
\arguments{
\item{referencia}{Codigo da referencia}

\item{cod_marca}{Codigo da marca}
}
\value{
Uma lista com duas tibbles. Na primeira tem modelos e codigos
na segunda tem codigos dos tipos de combustivel.
}
\description{
Codigo dos modelos
}
\examples{
# modelos da audia em 2022
get_modelos("292", "6")
}
