% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_fipe.R
\name{get_referencia}
\alias{get_referencia}
\title{Codigo do ano/mes de referencia}
\usage{
get_referencia(.ano)
}
\arguments{
\item{ano_}{}
}
\value{
Uma lista de
}
\description{
Codigo do ano/mes de referencia
}
\examples{
get_referencia(2015)
}
