% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_fipe.R
\name{pegar_preco}
\alias{pegar_preco}
\title{Preco do modelo na referencia}
\usage{
pegar_preco(link)
}
\arguments{
\item{link}{Link gerado em \code{\link{tabela_fipe}}}
}
\value{
Uma tibble com dado de preco do modelo na referencia
}
\description{
Preco do modelo na referencia
}
